// src/app/models/pessoa.model.ts
export class Pessoa {
    [x: string]: any;
    id?: number;
    nome?: string = '';
    nomeCompleto?: string = '';
    nomeFamilia?: string = '';
    grupo?: string = '';
    privilegio?: string = '';
    tipoPublicador?: string = '';
    dtNascimento?: Date;
    dtBatismo?: Date;
    sexo?: string = '';
    telefone?: string = '';
    endereco?: string = '';
  }
  